new version with rest and json, the json still is not like xml, new dev ahead

+++ poc_xcf/src/main/java/demo/restful/CategoryService.java
@@ -1,11 +1,13 @@
 package demo.restful;
 
-import org.springframework.beans.factory.annotation.Autowired;
-
 import javax.ws.rs.*;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.WebApplicationException;
+
+import org.apache.cxf.jaxrs.impl.ResponseBuilderImpl;
 
 @Path("/categoryservice")
-@Produces("application/xml")
+@Produces({"application/json","application/xml"})
 public class CategoryService {
     private CategoryDAO categoryDao = new CategoryDAO();
 
@@ -19,44 +21,79 @@ public class CategoryService {
 
     @GET
     @Path("/category/{id}")
+    @Produces({"application/json","application/xml"})
     public Category getCategory(@PathParam("id") String id) {
         System.out.println("getCategory called with category id:" + id);
-        return (Category) getCategoryDAO().getCategory(id);
+        Category cat = (Category) getCategoryDAO().getCategory(id);
+        if (cat == null) {
+            ResponseBuilderImpl builder = new ResponseBuilderImpl();
+            builder.status(Response.Status.BAD_REQUEST);
+            builder.type("application/xml");
+            builder.entity("<error>Category Not Found</error>");
+            throw new WebApplicationException(Response.Status.BAD_REQUEST);
+        } else {
+            return cat;
+        }
     }
 
     @POST
     @Path("/category")
-    @Consumes("application/xml")
-    public void addCategory(Category category) {
+    @Consumes({"application/json","application/xml"})
+    public Response addCategory(Category category) {
         System.out.println("addCategory called");
-        getCategoryDAO().addCategory(category);
+        Category cat = (Category) getCategoryDAO().getCategory(category.getCategoryId());
+        if (cat != null) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        } else {
+            getCategoryDAO().addCategory(category);
+            return Response.ok(category).build();
+        }
     }
 
     @DELETE
     @Path("/category/{id}")
-    public void deleteCategory(@PathParam("id") String id) {
+    public Response deleteCategory(@PathParam("id") String id) {
         System.out.println("deleteCategory called with id=" + id);
-        getCategoryDAO().deleteCategory(id);
+        Category cat = (Category) getCategoryDAO().getCategory(id);
+        if (cat == null) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        } else {
+            getCategoryDAO().deleteCategory(id);
+            return Response.ok().build();
+        }
     }
 
     @PUT
     @Path("/category")
-    public void updateCategory(Category category) {
+    public Response updateCategory(Category category) {
         System.out.println("updateCategory called with id=" + category.getCategoryId());
-        getCategoryDAO().updateCategory(category);
+        Category cat = (Category) getCategoryDAO().getCategory(category.getCategoryId());
+        if (cat == null) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        } else {
+            getCategoryDAO().updateCategory(category);
+            return Response.ok(category).build();
+        }
     }
 
     @POST
     @Path("category/book")
-    @Consumes("application/xml")
-    public void addBook(Category category) {
+    @Consumes({"application/json","application/xml"})
+    public Response addBook(Category category) {
         System.out.println("addBook with category id=" + category.getCategoryId());
-        getCategoryDAO().addBook(category);
+
+        Category cat = (Category) getCategoryDAO().getCategory(category.getCategoryId());
+        if (cat == null) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        } else {
+            getCategoryDAO().addBook(category);
+            return Response.ok(category).build();
+        }
     }
 
     @GET
     @Path("/category/{id}/books")
-    @Produces("application/xml")
+    @Produces({"application/json","application/xml"})
     public Category getBooks(@PathParam("id") String id) {
         System.out.println("getBooks called with id=" + id);
         Category cat = (Category) getCategoryDAO().getCategory(id);
